requirements.txt
numpy
pandas
torch
torchvision
flask
dlib
opencv-python
scikit-learn
matplotlib
jupyter
requests
tqdm
pillow
tensorflow
keras
urllib3
idna
certifi
numpy
gdown